+++ appliance/domain/Recipe/Cookbook/JobStart.php
@@ -44,6 +44,7 @@ use Teknoo\Recipe\RecipeInterface;
 use Teknoo\Space\Contracts\Recipe\Step\Job\CallNewJobInterface;
 use Teknoo\Space\Contracts\Recipe\Step\Job\NewJobNotifierInterface;
 use Teknoo\Space\Recipe\Step\Job\PrepareNewJobForm;
+use Teknoo\Space\Recipe\Step\NewJob\NewJobSetDefaults;
 use Teknoo\Space\Recipe\Step\PersistedVariable\LoadPersistedVariablesForJob;
 
 /**
@@ -65,6 +66,7 @@ class JobStart implements CookbookInterface
         private readonly LoadPersistedVariablesForJob $loadPersistedVariablesForJob,
         private readonly FormHandlingInterface $formHandling,
         private readonly FormProcessingInterface $formProcessing,
+        private readonly NewJobSetDefaults $newJobSetDefaults,
         private readonly NewJobNotifierInterface $newJobNotifier,
         private readonly JumpIf $jumpIf,
         private readonly CallNewJobInterface $callNewJob,
@@ -129,6 +131,8 @@ class JobStart implements CookbookInterface
         $recipe = $recipe->cook($this->formHandling, FormHandlingInterface::class, [], 40);
 
         $recipe = $recipe->cook($this->formProcessing, FormProcessingInterface::class, [], 50);
+
+        $recipe = $recipe->cook($this->newJobSetDefaults, NewJobSetDefaults::class, [], 60);
 
         $recipe = $recipe->cook($this->newJobNotifier, NewJobNotifierInterface::class, [], 70);