+++ appliance/domain/Recipe/Step/Job/JobSetDefaults.php
@@ -29,8 +29,8 @@ use DomainException;
 use Teknoo\East\Foundation\Manager\ManagerInterface;
 use Teknoo\East\Paas\Object\Cluster;
 use Teknoo\East\Paas\Object\Job;
-use Teknoo\Space\Object\Config\ClusterCatalog;
 use Teknoo\Space\Object\DTO\AccountWallet;
+use Teknoo\Space\Object\DTO\NewJob;
 
 /**
  * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
@@ -40,35 +40,37 @@ use Teknoo\Space\Object\DTO\AccountWallet;
  */
 class JobSetDefaults
 {
-    public function __construct(
-        private ClusterCatalog $catalog,
-    ) {
-    }
-
     public function __invoke(
         ManagerInterface $manager,
         Job $job,
         AccountWallet $accountWallet,
+        NewJob $newJob,
     ): self {
         $defaults = [];
         /**
          * @param Cluster[] $clusters
          */
-        $defaultsGenerator = function (iterable $clusters) use (&$defaults, $accountWallet): void {
+        $defaultsGenerator = function (iterable $clusters) use (&$defaults, $accountWallet, $newJob): void {
             foreach ($clusters as $cluster) {
-                $credential = $accountWallet[$cluster];
-                $config = $this->catalog->getCluster($cluster);
+                if ($cluster->isLocked()) {
+                    $credential = $accountWallet[$cluster];
 
-                $registryConfigName = $credential?->getRegistryConfigName();
-                if (empty($registryConfigName)) {
-                    throw new DomainException("Error, there are no registry config name for {$cluster}");
+                    $registryConfigName = $credential?->getRegistryConfigName();
+                    if (empty($registryConfigName)) {
+                        throw new DomainException("Error, there are no registry config name for {$cluster}");
+                    }
+
+                    $defaults = [
+                        'oci-registry-config-name' => $registryConfigName,
+                    ];
+
+                    $clusterName = (string) $cluster;
+                    if (isset($newJob->storageProvisionerPerCluster[$clusterName])) {
+                        $defaults['storage-provider'] = $newJob->storageProvisionerPerCluster[$clusterName];
+                    }
                 }
 
-                $defaults = [
-                    'oci-registry-config-name' => $registryConfigName,
-                    'storage-provider' => $config->storageProvisioner,
-                ];
-
+                //These default are currently shared for all clusters in the job, but not later
                 break;
             }
         };